+++ src/MeshData.h
@@ -14,7 +14,7 @@ public:
     std::vector<glm::vec3> vertices;
     std::vector<glm::vec3> normals;
     std::vector<glm::vec2> texCoords;
-    std::vector<std::vector<uint16_t>> indices;
+    std::vector<std::vector<uint32_t>> indices;
 
     static auto loadObj(const std::string &path) -> MeshData;
 };
